src/bigQuery/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-bigquery==1.22.0